requirements.txt
tweepy
requests
newspaper3k

